requirements.txt
python-telegram-bot
django
python-dotenv
requets